Entity Classes Integration
Modified all  entity, service, repository and controller classes.

- matching attribute names from the ERD.
- applied relationships and cardinality.
- MySql Database Connection (app properties)
- Code for autocreation of Tables to mysql database (application properties)

+++ backend/serbisyo/src/main/java/edu/cit/serbisyo/repository/ServiceCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import edu.cit.serbisyo.entity.ServiceCategoryEntity;
 
 @Repository
-public interface ServiceCategoryRepository extends JpaRepository<ServiceCategoryEntity, Integer> {
+public interface ServiceCategoryRepository extends JpaRepository<ServiceCategoryEntity, Long> {
 }
